RAG_React_tool/microservices/chroma/requirements.txt
PyPDF2
fastapi
langchain-community
chromadb==0.4.22
uvicorn
sentence-transformers
pypdf
numpy<2.0
